Revert "feat: day 7, 2024 - quick and dirty re-factoring of longs instead of ints everywhere"
This reverts commit 0b868bfb848ded14901b12cf3e495bb6a7884794.

+++ src/main/kotlin/eu/tinylinden/aoc/y2024/d01/Day01.kt
@@ -4,11 +4,11 @@ import kotlin.math.abs
 
 // https://adventofcode.com/2024/day/1
 
-fun distance(s: String): Long =
-    (left(s) zip right(s)).sumOf { (l, r) -> abs(l - r) }.toLong()
+fun distance(s: String): Int =
+    (left(s) zip right(s)).sumOf { (l, r) -> abs(l - r) }
 
-fun similarity(s: String): Long =
-    similarity(left(s), right(s).groupingBy { it }.eachCount()).toLong()
+fun similarity(s: String): Int =
+    similarity(left(s), right(s).groupingBy { it }.eachCount())
 
 private fun similarity(left: List<Int>, right: Map<Int, Int>): Int =
     left.sumOf { it * (right[it] ?: 0) }
